rename Paper to ts
